+++ src/test/java/roomescape/integration/AdminPageIntegrationTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.integration;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
